+++ changelog.md
@@ -24,6 +24,7 @@ Codion Change Log
 - TabbedDetailLayout.TabbedDetailController.selectParentTabs() bug fixed, now uses the correct indexOfComponent() instead of indexOfTabComponent().
 - TabbedApplicationLayout and TabbedDetailLayout now initialize panels using prepareComponent(), panels now contain focusable components when shown.
 - EntityPanel no longer triggers display request when activated.
+- EditAttributeDialogBuilder.editComponent() renamed component().
 
 ## 0.18.71
 ### is.codion.common.reactive